requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
dateparser==1.1.5
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
fredapi==0.5.0
frozendict==2.3.4
frozenlist==1.3.3
fsspec==2022.11.0
html5lib==1.1
icecream==2.1.3
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lightning-utilities==0.5.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.4
multitasking==0.0.11
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-binance==1.0.16
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==1.8.6
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.0
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboardX==2.5.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.13.1
torchmetrics==0.11.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.6.0
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
widgetsnbextension==4.0.5
yarl==1.8.2
yfinance==0.2.3
zipp==3.11.0
